factor app footer components into a single file
