+++ src/excalidraw-app/index.tsx
@@ -77,8 +77,7 @@ import { atom, Provider, useAtom } from "jotai";
 import { jotaiStore, useAtomWithInitialValue } from "../jotai";
 import { reconcileElements } from "./collab/reconciliation";
 import { parseLibraryTokensFromUrl, useHandleLibrary } from "../data/library";
-import EncryptedIcon from "../components/EncryptedIcon";
-import ExcalidrawPlusAppLink from "./components/ExcalidrawPlusAppLink";
+import { AppFooter } from "./components/AppFooter";
 
 polyfill();
 window.EXCALIDRAW_THROTTLE_RENDER = true;
@@ -641,10 +640,7 @@ const ExcalidrawWrapper = () => {
         theme={theme}
       >
         <Footer>
-          <div style={{ display: "flex", gap: ".5rem", alignItems: "center" }}>
-            <ExcalidrawPlusAppLink />
-            <EncryptedIcon />
-          </div>
+          <AppFooter />
         </Footer>
       </Excalidraw>
       {excalidrawAPI && <Collab excalidrawAPI={excalidrawAPI} />}
